backend/app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
python-dotenv==1.0.0
schedule==1.2.1
torch>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
dramatiq[redis]>=1.14.0
redis>=5.0.0
pandas>=2.0.0
openpyxl>=3.1.0  # 用于Excel导出
python-jose[cryptography]>=3.3.0  # 用于生成任务ID
aiofiles>=23.2.1  # 用于异步文件操作
psutil  # 添加 psutil 包用于系统资源监控
python-dotenv==1.0.0  # 添加 python-dotenv 支持